requirements.txt
pandas>=2.1.1
numpy>=1.24.4
pyyaml>=6.0